Lobby id still don't work, but I've added a ton of stuff

+++ Online/Events/OnlineEvent.cs
@@ -7,7 +7,7 @@ namespace RainMeadow
         public abstract EventTypeId eventType { get; } // serialized externally
         public OnlinePlayer from;// not serialized
         public OnlinePlayer to;// not serialized
-        public ulong eventId;
+        public ushort eventId;
         public TickReference dependsOnTick; // not serialized but universally supported, serialize if used for your event type
 
         public override string ToString()
@@ -15,7 +15,7 @@ namespace RainMeadow
             return $"{eventId}:{eventType}";
         }
 
-        public virtual long EstimatedSize { get => sizeof(ulong); }
+        public virtual long EstimatedSize { get => sizeof(ushort); }
 
         public virtual void CustomSerialize(Serializer serializer)
         {
